@@ -2,18 +2,18 @@ package com.cap.seattleemergencyhubs;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
 import android.util.Log;
 import android.view.View;
-import android.support.v4.view.GravityCompat;
-import android.support.v7.app.ActionBarDrawerToggle;
+import androidx.core.view.GravityCompat;
+import androidx.appcompat.app.ActionBarDrawerToggle;
 import android.view.MenuItem;
-import android.support.design.widget.NavigationView;
-import android.support.v4.widget.DrawerLayout;
+import com.google.android.material.navigation.NavigationView;
+import androidx.drawerlayout.widget.DrawerLayout;
 
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.Menu;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
@@ -117,7 +117,7 @@ public class MainActivity extends AppCompatActivity
             @Override
             public void onClick(View view) {
 
-                Intent trans = new Intent(MainActivity.this, NHUBSActivity.class);
+                Intent trans = new Intent(MainActivity.this, SelectedNeighborhoods.class);
                 trans.putExtra("transValue", nameTrans);
 
                 Log.wtf("myTag", "THIS LOG SHOWS VARIABLE BEFORE GOING TO NHUBSACTIVITY" );
@@ -151,6 +151,7 @@ public class MainActivity extends AppCompatActivity
             drawer.closeDrawer(GravityCompat.START);
         } else {
             super.onBackPressed();
+
         }
     }
 
@@ -187,10 +188,9 @@ public class MainActivity extends AppCompatActivity
         } else if (id == R.id.nav_about) {
 
         } else if (id == R.id.nav_neighborhood) {
-            Intent intent = new Intent(MainActivity.this, NHUBSActivity.class);
-            startActivity(intent);
 
-
+            Intent defaultNeighborhoods = new Intent(this, SelectedNeighborhoods.class);
+            startActivity(defaultNeighborhoods);
         } else if (id == R.id.nav_resources) {
             Intent intent = new Intent(MainActivity.this, ResoursesActivity.class);
             startActivity(intent);
